Finished the relationship tree service.
