+++ src/services/relationship_tree_service.py
@@ -1,7 +1,8 @@
 import logging
 from typing import Dict, List, Tuple, Any
 
-from PyQt6.QtGui import QStandardItemModel
+from PyQt6.QtCore import Qt
+from PyQt6.QtGui import QStandardItemModel, QStandardItem
 
 
 class RelationshipTreeService:
@@ -46,3 +47,70 @@ class RelationshipTreeService:
             parent_child_map[key].append((node_name, labels))
 
         return parent_child_map, skipped_records
+
+    def add_children(
+        self,
+        parent_name: str,
+        parent_item: QStandardItem,
+        path: List[str],
+        parent_child_map: Dict[Tuple[str, str, str], List[Tuple[str, List[str]]]],
+    ) -> None:
+        """
+        Add child nodes to the relationship tree with checkboxes.
+
+        Args:
+            parent_name (str): The name of the parent node.
+            parent_item (QStandardItem): The parent item in the tree.
+            path (List[str]): The path of node names to avoid cycles.
+            parent_child_map (dict): The parent-child map of relationships.
+        """
+        for (p_name, rel_type, direction), children in parent_child_map.items():
+            if p_name != parent_name:
+                continue
+
+            for child_name, child_labels in children:
+                if child_name in path:
+                    self.handle_cycles(parent_item, rel_type, direction, child_name)
+                    continue
+
+                arrow = "➡️" if direction == ">" else "⬅️"
+
+                # Create relationship item (non-checkable separator)
+                rel_item = QStandardItem(f"{arrow} [{rel_type}]")
+                rel_item.setFlags(
+                    Qt.ItemFlag.ItemIsEnabled | Qt.ItemFlag.ItemIsSelectable
+                )
+
+                # Create node item (checkable)
+                child_item = QStandardItem(
+                    f"🔹 {child_name} [{', '.join(child_labels)}]"
+                )
+                child_item.setData(child_name, Qt.ItemDataRole.UserRole)
+                child_item.setFlags(
+                    Qt.ItemFlag.ItemIsEnabled
+                    | Qt.ItemFlag.ItemIsSelectable
+                    | Qt.ItemFlag.ItemIsUserCheckable
+                )
+                child_item.setCheckState(Qt.CheckState.Unchecked)
+
+                rel_item.appendRow(child_item)
+                parent_item.appendRow(rel_item)
+
+                self.add_children(
+                    child_name, child_item, path + [child_name], parent_child_map
+                )
+
+    def handle_cycles(
+        self, parent_item: QStandardItem, rel_type: str, direction: str, child_name: str
+    ) -> None:
+        """
+        Handle cycles in the tree to prevent infinite loops.
+
+        Args:
+            parent_item (QStandardItem): The parent item.
+            rel_type (str): Relationship type.
+            direction (str): Relationship direction.
+            child_name (str): The child node involved in the cycle.
+        """
+        cycle_item = QStandardItem(f"🔁 Cycle: {child_name} ({rel_type}) [{direction}]")
+        parent_item.appendRow(cycle_item)
